requirements.txt
# local package
-e .

# external requirements
python-box==6.0.2
seaborn
pandas
numpy
scikit-learn
dvc
mlflow==2.2.2
tqdm
ensure
notebook
dagshub